Commenting

+++ controllers/event_controller.js
@@ -27,6 +27,7 @@ var createEvent = function (req, res) {
         errors: errors.errors
         });
     } else {
+        //storing each data to the variable
         let event = new Event();
         event.name = req.body.name;
         event.category = req.body.category;
@@ -47,6 +48,7 @@ var createEvent = function (req, res) {
                 //res.redirect('/');
             }
         });
+        //storing the a db created for event and user
         let userevents = new UserEvents();
         userevents.userid = req.user.id;
         userevents.eventid = event.id;
@@ -62,7 +64,7 @@ var createEvent = function (req, res) {
     }
 };
 
-//get single event
+//get single event in a page
 var getEvent = function(req, res){
     var days = ['Sunday','Monday','Tuesday','Wednesday','Thursday','Friday','Saturday'];
     var months = ['January','February','March','April','May','June','July','August','September','October','November','December'];
@@ -74,6 +76,8 @@ var getEvent = function(req, res){
         });
     });
 };
+
+//join an event letting the user data created to be stored and update the capcity
 var joinEvent = function(req,res){
     Event.findById(req.params.id, function(err, event) {
         if (err) {
@@ -87,12 +91,13 @@ var joinEvent = function(req,res){
             }
             else {
                 UserEvents.find({userid: req.user.id, eventid: event.id, type: "join"}, function (err, docs) {
+                    //check if user has previously register for the event
                     if (docs.length > 0){
-                        //req.flash('danger', 'already in event');
                         req.flash("danger", "already in event");
                         res.redirect('/');
                     }
                     else{
+                        //update the current event after joining
                         event.current_attendees = event.current_attendees + 1;
                         event.save(function (err) {
                             if (err) {
@@ -102,6 +107,8 @@ var joinEvent = function(req,res){
                                 req.flash('success', 'Join Succesful');
                             }
                         });
+
+                        //udpdate to the db of the event
                         let userevents = new UserEvents();
                         userevents.userid = req.user.id;
                         userevents.eventid = event.id;
@@ -118,6 +125,7 @@ var joinEvent = function(req,res){
             }
         }
          else {
+            //check if event is full
             req.flash('danger','Event is Full');
             res.redirect('/');
             return;
@@ -129,8 +137,8 @@ var joinEvent = function(req,res){
 var loadEvent =  function (req, res) {
     var id = req.params.id;
 
+    //find the data of current event
     Event.findById(id, function(err, event){
-
         if(event.creatorID != req.user.id){
             req.flash('danger', 'Not Authorized');
             return res.redirect('/');
@@ -143,7 +151,7 @@ var loadEvent =  function (req, res) {
     });
 };
 
-//post edit event
+//Update and save the event to the db
 var editEvent =  function (req, res) {
     var id = req.body.id;
     
@@ -174,13 +182,19 @@ var editEvent =  function (req, res) {
 //delete event
 var deleteEvent = function (req, res) {
     var id = req.body.id;
-
-    Event.findByIdAndRemove(id).exec();
-    req.flash('danger','Event Deleted');
-    res.redirect('/');
+    Event.findById(id, function (err, event) {
+        if(req.user.id == event.creatorID){
+            Event.findByIdAndRemove(id).exec();
+            req.flash('danger','Event Deleted');
+            res.redirect('/');
+        }else{
+            req.flash('danger','Not Authorized');
+            res.redirect('/event/'+id);
+        }
+    });
 };
 
-
+//validation use in the html entry
 var validate = (method) => {
     switch (method) {
         case 'saveEvent': {
